+++ example/a.ts
@@ -1,14 +1,20 @@
-import { Elysia, error, t } from '../src'
+import { Elysia } from '../src'
 import { req } from '../test/utils'
+import { newWebsocket, wsClosed, wsOpen } from '../test/ws/utils'
 
 const app = new Elysia()
-	.post(
-		'/login',
-		async function* ({ body: { username, password } }) {
-			for (let i = 0; i < 1000; i++) yield 'A\n'
-		},
-		{
-			body: t.Object({ username: t.String(), password: t.String() })
-		}
-	)
+	.ws('/ws', {
+		message() {}
+	})
+	.get('/ws', () => 'hi')
 	.listen(3000)
+
+console.log(app.fetch.toString())
+
+// const ws = newWebsocket(app.server!)
+
+// await wsOpen(ws)
+// await wsClosed(ws)
+
+const response = await app.handle(req('/ws')).then((x) => x.text())
+console.log(response) // .toBe('hi')