+++ src/adapter/web-standard/handler.ts
@@ -1,327 +1,21 @@
 /* eslint-disable sonarjs/no-nested-switch */
 /* eslint-disable sonarjs/no-duplicate-string */
-import { isNotEmpty, hasHeaderShorthand, StatusMap } from '../../utils'
+import {
+	createResponseHandler,
+	createStreamHandler,
+	handleFile,
+	handleSet,
+	responseToSetHeaders,
+	streamResponse
+} from '../utils'
 
-import { Cookie, serializeCookie } from '../../cookies'
+import { ElysiaFile } from '../../universal/file'
+import { isNotEmpty } from '../../utils'
+import { Cookie } from '../../cookies'
+import { ElysiaCustomStatusResponse } from '../../error'
 
 import type { Context } from '../../context'
 import type { AnyLocalHook } from '../../types'
-import { ElysiaCustomStatusResponse } from '../../error'
-import { ElysiaFile } from '../../universal/file'
-
-// type SetResponse = Omit<Context['set'], 'status'> & {
-// 	status: number
-// }
-
-const handleFile = (response: File | Blob, set?: Context['set']) => {
-	const size = response.size
-
-	if (
-		(!set && size) ||
-		(size &&
-			set &&
-			set.status !== 206 &&
-			set.status !== 304 &&
-			set.status !== 412 &&
-			set.status !== 416)
-	) {
-		if (set) {
-			if (set.headers instanceof Headers) {
-				let setHeaders: Record<string, any> = {
-					'accept-ranges': 'bytes',
-					'content-range': `bytes 0-${size - 1}/${size}`,
-					'transfer-encoding': 'chunked'
-				}
-
-				if (hasHeaderShorthand)
-					setHeaders = (set.headers as unknown as Headers).toJSON()
-				else {
-					setHeaders = {}
-					for (const [key, value] of set.headers.entries())
-						if (key in set.headers) setHeaders[key] = value
-				}
-
-				return new Response(response as Blob, {
-					status: set.status as number,
-					headers: setHeaders
-				})
-			}
-
-			if (isNotEmpty(set.headers))
-				return new Response(response as Blob, {
-					status: set.status as number,
-					headers: Object.assign(
-						{
-							'accept-ranges': 'bytes',
-							'content-range': `bytes 0-${size - 1}/${size}`,
-							'transfer-encoding': 'chunked'
-						} as any,
-						set.headers
-					)
-				})
-		}
-
-		return new Response(response as Blob, {
-			headers: {
-				'accept-ranges': 'bytes',
-				'content-range': `bytes 0-${size - 1}/${size}`,
-				'transfer-encoding': 'chunked'
-			}
-		})
-	}
-
-	return new Response(response as Blob)
-}
-
-export const parseSetCookies = (headers: Headers, setCookie: string[]) => {
-	if (!headers) return headers
-
-	headers.delete('set-cookie')
-
-	for (let i = 0; i < setCookie.length; i++) {
-		const index = setCookie[i].indexOf('=')
-
-		headers.append(
-			'set-cookie',
-			`${setCookie[i].slice(0, index)}=${
-				setCookie[i].slice(index + 1) || ''
-			}`
-		)
-	}
-
-	return headers
-}
-
-const responseToSetHeaders = (response: Response, set?: Context['set']) => {
-	if (set?.headers) {
-		if (response) {
-			if (hasHeaderShorthand)
-				// @ts-expect-error
-				Object.assign(set.headers, response.headers.toJSON())
-			else
-				for (const [key, value] of response.headers.entries())
-					if (key in set.headers) set.headers[key] = value
-		}
-
-		if (set.status === 200) set.status = response.status
-
-		// ? `content-encoding` prevent response streaming
-		if (set.headers['content-encoding'])
-			delete set.headers['content-encoding']
-
-		return set
-	}
-
-	if (!response)
-		return {
-			headers: {},
-			status: set?.status ?? 200
-		}
-
-	if (hasHeaderShorthand) {
-		set = {
-			// @ts-expect-error
-			headers: response.headers.toJSON(),
-			status: set?.status ?? 200
-		}
-
-		// ? `content-encoding` prevent response streaming
-		if (set.headers['content-encoding'])
-			delete set.headers['content-encoding']
-
-		return set
-	}
-
-	set = {
-		headers: {},
-		status: set?.status ?? 200
-	}
-
-	for (const [key, value] of response.headers.entries()) {
-		// ? `content-encoding` prevent response streaming
-
-		if (key === 'content-encoding') continue
-
-		if (key in set.headers) set.headers[key] = value
-	}
-
-	return set
-}
-
-const handleStream = async (
-	generator: Generator | AsyncGenerator,
-	set?: Context['set'],
-	request?: Request
-) => {
-	let init = generator.next()
-	if (init instanceof Promise) init = await init
-
-	if (init.done) {
-		if (set) return mapResponse(init.value, set, request)
-		return mapCompactResponse(init.value, request)
-	}
-
-	if (set?.headers) {
-		if (!set.headers['transfer-encoding'])
-			set.headers['transfer-encoding'] = 'chunked'
-		if (!set.headers['content-type'])
-			set.headers['content-type'] = 'text/event-stream; charset=utf-8'
-	} else {
-		set = {
-			status: 200,
-			headers: {
-				'content-type': 'text/event-stream; charset=utf-8',
-				'transfer-encoding': 'chunked'
-			}
-		}
-	}
-
-	return new Response(
-		new ReadableStream({
-			async start(controller) {
-				let end = false
-
-				request?.signal?.addEventListener('abort', () => {
-					end = true
-
-					try {
-						controller.close()
-					} catch {
-						// nothing
-					}
-				})
-
-				if (init.value !== undefined && init.value !== null) {
-					if (typeof init.value === 'object')
-						try {
-							controller.enqueue(
-								// @ts-expect-error this is a valid operation
-								Buffer.from(JSON.stringify(init.value))
-							)
-						} catch {
-							controller.enqueue(
-								// @ts-expect-error this is a valid operation
-								Buffer.from(init.value.toString())
-							)
-						}
-					else
-						controller.enqueue(
-							// @ts-expect-error this is a valid operation
-							Buffer.from(init.value.toString())
-						)
-				}
-
-				for await (const chunk of generator) {
-					if (end) break
-					if (chunk === undefined || chunk === null) continue
-
-					if (typeof chunk === 'object')
-						try {
-							controller.enqueue(
-								// @ts-expect-error this is a valid operation
-								Buffer.from(JSON.stringify(chunk))
-							)
-						} catch {
-							controller.enqueue(
-								// @ts-expect-error this is a valid operation
-								Buffer.from(chunk.toString())
-							)
-						}
-					else
-						controller.enqueue(
-							// @ts-expect-error this is a valid operation
-							Buffer.from(chunk.toString())
-						)
-
-					// Wait for the next event loop
-					// Otherwise the data will be mixed up
-					await new Promise<void>((resolve) =>
-						setTimeout(() => resolve(), 0)
-					)
-				}
-
-				try {
-					controller.close()
-				} catch {
-					// nothing
-				}
-			}
-		}),
-		set as any
-	)
-}
-
-export async function* streamResponse(response: Response) {
-	const body = response.body
-
-	if (!body) return
-
-	const reader = body.getReader()
-	const decoder = new TextDecoder()
-
-	try {
-		while (true) {
-			const { done, value } = await reader.read()
-			if (done) break
-
-			yield decoder.decode(value)
-		}
-	} finally {
-		reader.releaseLock()
-	}
-}
-
-export const handleSet = (set: Context['set']) => {
-	if (typeof set.status === 'string') set.status = StatusMap[set.status]
-
-	if (set.cookie && isNotEmpty(set.cookie)) {
-		const cookie = serializeCookie(set.cookie)
-
-		if (cookie) set.headers['set-cookie'] = cookie
-	}
-
-	if (set.headers['set-cookie'] && Array.isArray(set.headers['set-cookie'])) {
-		set.headers = parseSetCookies(
-			new Headers(set.headers as any) as Headers,
-			set.headers['set-cookie']
-		) as any
-	}
-}
-
-export const mergeResponseWithSetHeaders = (
-	response: Response,
-	set: Context['set']
-) => {
-	if (
-		(response as Response).status !== set.status &&
-		set.status !== 200 &&
-		((response.status as number) <= 300 ||
-			(response.status as number) > 400)
-	)
-		response = new Response(response.body, {
-			headers: response.headers,
-			status: set.status as number
-		})
-
-	let isCookieSet = false
-
-	if (set.headers instanceof Headers)
-		for (const key of set.headers.keys()) {
-			if (key === 'set-cookie') {
-				if (isCookieSet) continue
-
-				isCookieSet = true
-
-				for (const cookie of set.headers.getSetCookie())
-					response.headers.append('set-cookie', cookie)
-			} else response.headers.append(key, set.headers?.get(key) ?? '')
-		}
-	else
-		for (const key in set.headers)
-			(response as Response).headers.append(key, set.headers[key] as any)
-
-	return response
-}
 
 export const mapResponse = (
 	response: unknown,
@@ -384,23 +78,7 @@ export const mapResponse = (
 				return Response.json(response, set as any)
 
 			case 'Response':
-				response = mergeResponseWithSetHeaders(
-					response as Response,
-					set
-				)
-
-				if (
-					!(response as Response).headers.has('content-length') &&
-					(response as Response).headers.get('transfer-encoding') ===
-						'chunked'
-				)
-					return handleStream(
-						streamResponse(response as Response),
-						responseToSetHeaders(response as Response, set),
-						request
-					) as any
-
-				return response as Response
+				return handleResponse(response as Response, set, request)
 
 			case 'Error':
 				return errorToResponse(response as Error, set)
@@ -430,25 +108,9 @@ export const mapResponse = (
 				return new Response(response as FormData, set as any)
 
 			default:
-				if (response instanceof Response) {
-					response = mergeResponseWithSetHeaders(
-						response as Response,
-						set
-					)
-
-					if (
-						(response as Response).headers.get(
-							'transfer-encoding'
-						) === 'chunked'
-					)
-						return handleStream(
-							streamResponse(response as Response),
-							responseToSetHeaders(response as Response, set),
-							request
-						) as any
-
-					return response as Response
-				}
+				// recheck Response, Promise, Error because some library may extends Response
+				if (response instanceof Response)
+					return handleResponse(response as Response, set, request)
 
 				if (response instanceof Promise)
 					return response.then((x) => mapResponse(x, set)) as any
@@ -584,23 +246,7 @@ export const mapEarlyResponse = (
 				return Response.json(response, set as any)
 
 			case 'Response':
-				response = mergeResponseWithSetHeaders(
-					response as Response,
-					set
-				)
-
-				if (
-					!(response as Response).headers.has('content-length') &&
-					(response as Response).headers.get('transfer-encoding') ===
-						'chunked'
-				)
-					return handleStream(
-						streamResponse(response as Response),
-						responseToSetHeaders(response as Response, set),
-						request
-					) as any
-
-				return response as Response
+				return handleResponse(response as Response, set, request)
 
 			case 'Promise':
 				// @ts-ignore
@@ -631,25 +277,8 @@ export const mapEarlyResponse = (
 				return new Response(response?.toString(), set as any)
 
 			default:
-				if (response instanceof Response) {
-					response = mergeResponseWithSetHeaders(
-						response as Response,
-						set
-					)
-
-					if (
-						(response as Response).headers.get(
-							'transfer-encoding'
-						) === 'chunked'
-					)
-						return handleStream(
-							streamResponse(response as Response),
-							responseToSetHeaders(response as Response, set),
-							request
-						) as any
-
-					return response as Response
-				}
+				if (response instanceof Response)
+					return handleResponse(response, set, request)
 
 				if (response instanceof Promise)
 					return response.then((x) => mapEarlyResponse(x, set)) as any
@@ -1014,3 +643,13 @@ export const createStaticHandler = (
 	)
 		return response.clone.bind(response)
 }
+
+const handleResponse = createResponseHandler({
+	mapResponse,
+	mapCompactResponse
+})
+
+const handleStream = createStreamHandler({
+	mapResponse,
+	mapCompactResponse
+})