+++ test/adapter/web-standard/map-compact-response.test.ts
@@ -139,7 +139,7 @@ describe('Web Standard - Map Compact Response', () => {
 	})
 
 	it('map Response and merge Headers', async () => {
-		const response = mapCompactResponse(
+		const response = await mapCompactResponse(
 			new Response('Shiroko', {
 				headers: {
 					Name: 'Himari'