+++ test/adapter/web-standard/map-early-response.test.ts
@@ -201,7 +201,7 @@ describe('Web Standard - Map Early Response', () => {
 	})
 
 	it('map Response with custom context', async () => {
-		const response = mapEarlyResponse(new Response('Shiroko'), context)
+		const response = await mapEarlyResponse(new Response('Shiroko'), context)
 		const headers = response?.headers.toJSON()
 
 		expect(response).toBeInstanceOf(Response)
@@ -210,7 +210,7 @@ describe('Web Standard - Map Early Response', () => {
 	})
 
 	it('map Response and merge Headers', async () => {
-		const response = mapEarlyResponse(
+		const response = await mapEarlyResponse(
 			new Response('Shiroko', {
 				headers: {
 					Name: 'Himari'